The Cricket-Match evaluation window
It pops up a dialog window for the match evaluation of the cricket game. Evaluating the scores of the respective teams specified in the Combobox of the team's section. 
It is connected with the cricket_match database using the Python SQLite Library and the interface is designed using the Qt Designer and PyQt Library along with the pyuic5 utility.